+++ R/utils_api-with.R
@@ -28,3 +28,37 @@ with.FutureStrategyList <- function(data, expr, ...) {
   
   expr
 }
+
+
+#' @param envir The environment where the future plan should be set.
+#'
+#' @rdname plan
+#' @export
+localPlan <- function(strategy = NULL, ..., .cleanup = NA, substitute = TRUE, envir = parent.frame()) {
+  if (substitute) strategy <- substitute(strategy)
+  if (identical(envir, topenv())) {
+    stop("Can only set a local future plan inside a function or inside local()")
+  }
+
+  ## Set up on.exit() in the parent frame
+  oplan <- plan("list")
+  undoPlan <- function() plan(oplan, .cleanup = TRUE)
+  call <- as.call(list(undoPlan))
+  args <- list(call, add = TRUE, after = TRUE)
+  do.call(base::on.exit, args = args, envir = envir)
+  
+  ## Set the new, temporary plan, without cleaning up the current one
+  plan(strategy, substitute = FALSE, .cleanup = FALSE, ...)
+  
+  invisible(oplan)
+}
+
+
+ping <- function(..., envir = parent.frame()) {
+  a <- 42
+  
+  ## Set up on.exit() in the parent frame
+  thunk <- as.call(list(function() print(2*a)))
+  args <- list(thunk, add = TRUE, after = TRUE)
+  do.call(base::on.exit, args = args, envir = envir)
+}